+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/DownloadStateReducer.kt
@@ -25,6 +25,8 @@ internal object DownloadStateReducer {
             is DownloadAction.RemoveAllDownloadsAction -> {
                 state.copy(downloads = emptyMap())
             }
+            is DownloadAction.RestoreDownloadsStateAction -> state
+            is DownloadAction.RestoreDownloadStateAction -> updateDownloads(state, action.download)
         }
     }