+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/AbstractFetchDownloadService.kt
@@ -68,6 +68,7 @@ import mozilla.components.feature.downloads.facts.emitNotificationResumeFact
 import mozilla.components.feature.downloads.facts.emitNotificationTryAgainFact
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.ktx.kotlin.sanitizeURL
+import mozilla.components.support.ktx.kotlinx.coroutines.throttleLatest
 import mozilla.components.support.utils.DownloadUtils
 import java.io.File
 import java.io.FileOutputStream
@@ -100,7 +101,7 @@ abstract class AbstractFetchDownloadService : Service() {
     internal var compatForegroundNotificationId: Int = COMPAT_DEFAULT_FOREGROUND_ID
     private val logger = Logger("AbstractFetchDownloadService")
 
-    internal var downloadJobs = mutableMapOf<Long, DownloadJobState>()
+    internal var downloadJobs = mutableMapOf<String, DownloadJobState>()
 
     // TODO Move this to browser store and make immutable:
     // https://github.com/mozilla-mobile/android-components/issues/7050
@@ -156,7 +157,7 @@ abstract class AbstractFetchDownloadService : Service() {
             @Suppress("LongMethod")
             override fun onReceive(context: Context, intent: Intent?) {
                 val downloadId =
-                    intent?.extras?.getLong(DownloadNotification.EXTRA_DOWNLOAD_ID) ?: return
+                    intent?.extras?.getString(DownloadNotification.EXTRA_DOWNLOAD_ID) ?: return
                 val currentDownloadJobState = downloadJobs[downloadId] ?: return
 
                 when (intent.action) {
@@ -246,24 +247,28 @@ abstract class AbstractFetchDownloadService : Service() {
     }
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        val download = intent?.getLongExtra(EXTRA_DOWNLOAD_ID, -1)?.let {
+        val download = intent?.getStringExtra(EXTRA_DOWNLOAD_ID)?.let {
             store.state.downloads[it]
         } ?: return START_REDELIVER_INTENT
 
         // If the job already exists, then don't create a new ID. This can happen when calling tryAgain
         val foregroundServiceId = downloadJobs[download.id]?.foregroundServiceId ?: Random.nextInt()
 
+        val actualStatus = if (download.status == INITIATED) DOWNLOADING else download.status
+
         // Create a new job and add it, with its downloadState to the map
         val downloadJobState = DownloadJobState(
-            state = download.copy(status = DOWNLOADING),
+            state = download.copy(status = actualStatus),
             foregroundServiceId = foregroundServiceId,
-            status = DOWNLOADING
+            status = actualStatus
         )
 
         store.dispatch(DownloadAction.UpdateDownloadAction(downloadJobState.state))
 
-        downloadJobState.job = CoroutineScope(IO).launch {
-            startDownloadJob(downloadJobState)
+        if (actualStatus == DOWNLOADING) {
+            downloadJobState.job = CoroutineScope(IO).launch {
+                startDownloadJob(downloadJobState)
+            }
         }
 
         downloadJobs[download.id] = downloadJobState
@@ -625,7 +630,15 @@ abstract class AbstractFetchDownloadService : Service() {
     internal fun copyInChunks(downloadJobState: DownloadJobState, inStream: InputStream, outStream: OutputStream) {
         val data = ByteArray(CHUNK_SIZE)
         logger.debug("starting copyInChunks ${downloadJobState.state.url}" +
-                " currentBytesCopied ${downloadJobState.currentBytesCopied}")
+                " currentBytesCopied ${downloadJobState.state.currentBytesCopied}")
+
+        val throttleUpdateDownload = throttleLatest<Long>(
+            PROGRESS_UPDATE_INTERVAL,
+            coroutineScope = CoroutineScope(IO)
+        ) { copiedBytes ->
+            val newState = downloadJobState.state.copy(currentBytesCopied = copiedBytes)
+            updateDownloadState(newState)
+        }
 
         // To ensure that we copy all files (even ones that don't have fileSize, we must NOT check < fileSize
         while (getDownloadJobStatus(downloadJobState) == DOWNLOADING) {
@@ -635,10 +648,7 @@ abstract class AbstractFetchDownloadService : Service() {
             if (bytesRead == -1) { break }
             downloadJobState.currentBytesCopied += bytesRead
 
-            val newState = downloadJobState.state.copy(
-                currentBytesCopied = downloadJobState.currentBytesCopied
-            )
-            updateDownloadState(newState)
+            throttleUpdateDownload(downloadJobState.currentBytesCopied)
 
             outStream.write(data, 0, bytesRead)
         }