+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/DownloadsFeature.kt
@@ -136,7 +136,7 @@ class DownloadsFeature(
      * Calls the tryAgain function of the corresponding [DownloadManager]
      */
     @Suppress("Unused")
-    fun tryAgain(id: Long) {
+    fun tryAgain(id: String) {
         downloadManager.tryAgain(id)
     }